:fire: refactor structure and file names of view
