+++ src/view/component/Logout/Logout.presentational.js
@@ -1,21 +1,8 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { Redirect } from 'react-router-dom'
-import { connect } from 'react-redux'
-import logoutAction from '../redux/action/logout'
 
-const mapStateToProps = state => (state.user)
-
-const mapDispatchToProps = (dispatch) => ({
-  logout: async () => dispatch(logoutAction())
-})
-
-@connect(mapStateToProps, mapDispatchToProps)
 class Logout extends Component {
-  static propTypes=({
-    logout: PropTypes.func,
-    location: PropTypes.object
-  })
   async _signOut () {
     await this.props.logout()
   }
@@ -27,6 +14,11 @@ class Logout extends Component {
       <Redirect to={{pathname: '/', state: { from: this.props.location }}} />
     )
   }
+}
+
+Logout.propTypes = {
+  logout: PropTypes.func,
+  location: PropTypes.object
 }
 
 export default Logout
