+++ src/view/component/Preview/index.less
