+++ src/view/component/Register/Register.presentational.js
