+++ src/view/component/Setting/Detail.connected.js
