+++ src/view/component/Setting/Setting.connected.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux'
-import Setting from '../../component/Setting'
+import Setting from './Setting.presentational'
 
 const mapStateToProps = (state) => ({
   userInfo: state.user.userInfo
