+++ src/view/component/User/User.presentational.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import CSSModules from 'react-css-modules'
 // import { NavLink } from 'react-router-dom'
-import Header from '../../layout/Header'
-import Footer from '../../layout/Footer/'
+import Header from '../../layout/Header/Header.connected'
+import Footer from '../../layout/Footer/Footer.presentational'
 import styles from './index.less'
 
 @CSSModules(styles)
