+++ src/view/component/Welcome/Welcome.connected.js
@@ -1,7 +1,7 @@
-import Welcome from '../component/Welcome'
+import Welcome from './Welcome.presentational'
 import { connect } from 'react-redux'
 import { withRouter } from 'react-router-dom'
-import { checkLoginRedirect } from '../common/util/redirect'
+import { checkLoginRedirect } from '../../common/util/redirect'
 
 const mapStateToProps = (state) => {
   return {
