+++ src/view/component/WelcomeSection/WelcomeSection.presentational.js
