+++ src/view/layout/Footer/Footer.presentational.js
