+++ src/view/layout/Header/Header.presentational.js
@@ -2,35 +2,14 @@ import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom'
 import CSSModules from 'react-css-modules'
-// import classNames from 'classnames'
-import { connect } from 'react-redux'
 import Avatar from '_common_ui/Avatar'
 import Button from '_common_ui/Button'
 import Icon from '_common_ui/Icon'
 import { Modal, ModalHeader, ModalBody } from '_common_ui/Modal'
-import Login from '../../component/Login/'
-import Register from '../../component/Register/'
-import { login as loginAction } from '../../redux/action/login'
-import { register as registerAction } from '../../redux/action/register'
-import {
-  update as updateAction,
-  updateAvatarUnlogined as updateAvatarUnloginedAction
-} from '../../redux/action/user'
+import Login from '../../component/Login/Login.presentational'
+import Register from '../../component/Register/Register.presentational'
 import styles from './index.less'
 
-const mapStateToProps = (state) => ({
-  otherInfo: state.user.otherInfo,
-  flashMessage: state.flashMessage
-})
-
-const mapDispatchToProps = (dispatch) => ({
-  login: async (user) => dispatch(loginAction(user)),
-  register: async (user) => dispatch(registerAction(user)),
-  update: async (info) => dispatch(updateAction(info)),
-  updateAvatarUnlogined: (result) => dispatch(updateAvatarUnloginedAction(result))
-})
-
-@connect(mapStateToProps, mapDispatchToProps)
 @CSSModules(styles)
 class Header extends PureComponent {
   constructor (props) {
@@ -66,22 +45,22 @@ class Header extends PureComponent {
             <a href="http://muwenzi.com"><li>旧版</li></a>
           </div>
           { avatarSrc
-          ? <div styleName="right">
-            <Icon type="message" styleName="icon-message" />
-            <div styleName="dropdown-wrapper">
-              <Avatar shape="circle"
-                      size="default"
-                      src={avatarSrc}
-                      styleName="avatar"
-              />
-              <ul styleName="dropdown-container">
-                <Link to="/user"><li>我的主页</li></Link>
-                <Link to="/settings/profile"><li>设置</li></Link>
-                <Link to="/logout"><li>退出</li></Link>
-              </ul>
+            ? <div styleName="right">
+              <Icon type="message" styleName="icon-message" />
+              <div styleName="dropdown-wrapper">
+                <Avatar shape="circle"
+                        size="default"
+                        src={avatarSrc}
+                        styleName="avatar"
+                />
+                <ul styleName="dropdown-container">
+                  <Link to="/user"><li>我的主页</li></Link>
+                  <Link to="/settings/profile"><li>设置</li></Link>
+                  <Link to="/logout"><li>退出</li></Link>
+                </ul>
+              </div>
+              <Link to={buttonLink}><Button size="sm" ghost>{buttonName}</Button></Link>
             </div>
-            <Link to={buttonLink}><Button size="sm" ghost>{buttonName}</Button></Link>
-          </div>
             : <div styleName="right">
               <div styleName="nav-btn" onClick={() => this._handleModal('login')}>登录</div>
               <Button onClick={() => this._handleModal('register')}>{buttonName}</Button>
@@ -107,7 +86,7 @@ class Header extends PureComponent {
                 </ModalBody>
               </Modal>
             </div>
-            }
+          }
         </div>
       </header>
     )
