+++ src/view/redux/action/register.action.js
@@ -1,5 +1,5 @@
 import constActionType from '../constant/actionType'
-import { showFlashMessage } from '../action/flashMessage'
+import { showFlashMessage } from './flashMessage.action'
 
 const startRegister = () => ({
   type: constActionType.REGISTER_STARTED
