+++ src/view/redux/reducer/preview.reducer.js
